backend/requirements.txt
fastapi
uvicorn
duckdb
sentence-transformers
faiss-cpu
networkx
python-louvain
bertopic
google-generativeai
pandas
pyarrow
pydantic
python-dotenv
nomic
